+++ src/features/ChatItem/components/Avatar.tsx
