+++ src/features/ChatItem/components/BorderSpacing.tsx
