+++ src/features/ChatItem/components/ErrorContent.tsx
